Change rest controller
