+++ src/main/java/ru/projects/calories/controller/api/ProductsRestController.java
@@ -1,8 +1,5 @@
 package ru.projects.calories.controller.api;
 
-import ru.projects.calories.model.Dish;
-import ru.projects.calories.model.Product;
-import ru.projects.calories.service.DishService;
 import ru.projects.calories.service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -10,28 +7,24 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import ru.projects.calories.model.Product;
 
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/v1/calories")
-public class CaloriesRestController
+@RequestMapping("/api/v1/products")
+public class ProductsRestController
 {
 	private final ProductService productService;
-	private final DishService dishService;
 
 	@Autowired
-	public CaloriesRestController(ProductService productService, DishService dishService)
+	public ProductsRestController(ProductService productService)
 	{
 		this.productService = productService;
-		this.dishService = dishService;
 	}
 
-	/**
-	 * Products restful api
-	 */
 
-	@GetMapping({ "/products", "/products/" })
+	@GetMapping({ "", "/" })
 	public ResponseEntity<List<Product>> getProducts()
 	{
 		List<Product> products = productService.getAll();
@@ -39,31 +32,11 @@ public class CaloriesRestController
 		return ResponseEntity.ok(products);
 	}
 
-	@GetMapping("/products/{name}")
+	@GetMapping("/{name}")
 	public ResponseEntity<Product> getProduct(@PathVariable String name)
 	{
 		Product product = productService.getProductByName(name);
 
 		return ResponseEntity.ok(product);
-	}
-
-	/**
-	 * Dishes restful api
-	 */
-
-	@GetMapping({"/dishes", "/dishes/"})
-	public ResponseEntity<List<Dish>> getDishes()
-	{
-		List<Dish> dishes = this.dishService.getAll();
-
-		return ResponseEntity.ok(dishes);
-	}
-
-	@GetMapping("/dishes/{name}")
-	public ResponseEntity<Dish> getDish(@PathVariable String name)
-	{
-		Dish dish = dishService.getDishByName(name);
-
-		return ResponseEntity.ok(dish);
 	}
 }
