+++ src/main/java/ru/projects/calories/controller/api/UserRestController.java
@@ -1,0 +1,10 @@
+package ru.projects.calories.controller.api;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/v1/users")
+public class UserRestController
+{
+}
